Allow renaming states
Setting meaningful names for states greatly helps understanding what the
automaton does. This commit allows the user to set display names for the
states (the internal ID being the default), while keeping the internal
IDs intact. It also adjusts the layout to account for longer user-
defined names: the states may be wider than 100px if needed (but never
narrower). It uses flexbox, which has pretty good browser support
<https://caniuse.com/flexbox>, and degrades gracefully even in browsers
not supporting it.